+++ src/exercise_3/infrastructure/index.ts
@@ -1,4 +1,2 @@
-export * from "./credit-card-payment.processor"
-export * from "./new-payment.processor"
-export * from "./pay-pal-payment.processor"
-export * from "./adapters/new-payment.adapter"
+export * from "./adapters/new-payment.adapter";
+export * from "./repositories/";
