+++ src/exercise_3/infrastructure/repositories/credit-card-payment.processor.ts
@@ -1,5 +1,5 @@
-import { MessagesConfig } from '../config';
-import { PaymentProcessor } from '../domain';
+import { MessagesConfig } from '../../config';
+import { PaymentProcessor } from '../../domain';
 
 export class CreditCardPaymentProcessor implements PaymentProcessor {
     processPayment(account: string, amount: number): void {
