+++ src/exercise_3/infrastructure/repositories/index.ts
@@ -1,0 +1,3 @@
+export * from "./credit-card-payment.processor"
+export * from "./new-payment.system"
+export * from "./pay-pal-payment.processor"
